node: dedupe observability boot-status log into a shared helper
The 8-line "Observability enabled / agentId unset → warn" block was byte-for-byte
duplicated in both agent_server.ts and audio_stream_server.ts (the Python side
already shares it via _server_base._log_observability_status). Extract
logObservabilityStatus(agentId) into observability.ts — the module both servers
already import — so the user-facing warning has a single source of truth.
