+++ crates/agent-transport-node/adapters/livekit/agent_server.ts
@@ -29,7 +29,7 @@ import { mkdirSync } from 'node:fs';
 import { SipEndpoint } from 'agent-transport';
 import { initializeLogger, InferenceRunner, runWithJobContext, log as agentLog } from '@livekit/agents';
 import { JobContext } from './session_context.js';
-import { getObservabilityUrl } from './observability.js';
+import { logObservabilityStatus } from './observability.js';
 import { finalizeSession } from './_session_finalize.js';
 import { runServerCleanup, forceShutdownAgentSession, installUnhandledRejectionHandler, registerSignalCleanup } from './_session_teardown.js';
 import { brokerFor, isAudioEvent } from './_audio_events.js';
@@ -355,16 +355,7 @@ export class AgentServer {
     this.startHttpServer();
     console.log(`HTTP server on http://${this.host}:${this.port}`);
 
-    const obsUrl = getObservabilityUrl();
-    if (obsUrl && this.agentId) {
-      console.log(`Observability enabled, target ${obsUrl}`);
-    } else if (obsUrl) {
-      console.warn(
-        `Observability is configured (AGENT_OBSERVABILITY_URL=${obsUrl}) but agentId is ` +
-          `unset — session reports will NOT be uploaded. Pass agentId to the server ` +
-          `constructor or set the AGENT_ID env var to enable observability.`,
-      );
-    }
+    logObservabilityStatus(this.agentId);
 
     // Start SIP event loop. Track the promise so we can await its exit
     // during shutdown — without this the infinite while loop would pin
